feat: create booking-summary page
Here are the changes:
1. Create booking-summary page. This is the page that user will open
after they choose a car to book. Creating this page is necessary in
order to match the design that the team has created.
2. Upload car image as dummy. The image is uploaded to /public/assets.
This car image is used as dummy to help development process. In next
iteration, this image may be changed to its actual image.
3. Delete Inter classname in RootLayout. The reason behind this decision
is because we also use another font beside Inter.

Here are the limitations:
1. Fonts that are used in this commit haven't matched the design.
2. There is no header and footer.
3. Data such as car name aren't fetched from database, instead they are
written statically.
4. There are no onClick functions in buttons.
5. There are minor styling mismatches.
In order to fix these limitations, further developments are needed.