fix: navbar for logged in and logged out user
Here are the changes:
1. Before this commit, the search-car page doesn't render the correct
navbar. For example, even if the user is logged in, the page will still
render the logged out version of the navbar. In this commit, the page
should render the correct navbar for logged in and logged out users.
2. I need to change the SearchContextFunction to SearchContextFunctionx
in order to avoid the same name. Thus, I also need to change the pages
that are affected by this change.