fix: return to search-car page from booking-sum
Before this commit, when user wanted to go back to the search-car page
after going to booking-summary page, the search-car page used the
default value of booking date for search-car. This is a bad ux. In this
commit, I use the start and date params in booking-summary page to be
used as params in search-car page. Thus, the search-car page uses the
same date as previously choosen by the user.