fix: month and time issue when creating rent doc
Before this commit, when a user confirmed a booking, the rent document
that were created didn't have the correct month and time. The wrong
month was submitted because month, in ISO format, start at zero instead
of one. In order to fix this issue, the month should be decreased by
one. Meanwhile, the wrong time was used because mongodb uses GMT by
default while the users don't live in places where GMT are used. I use 7
as the default hour. Keep in mind, it is a hotfix. It will only resolve
the issue for users who live in GMT+7 zone.